+++ sharding-proxy/src/main/java/io/shardingsphere/proxy/backend/common/StatementExecuteBackendHandler.java
@@ -17,7 +17,6 @@
 
 package io.shardingsphere.proxy.backend.common;
 
-import com.sun.rowset.CachedRowSetImpl;
 import io.shardingsphere.core.constant.DatabaseType;
 import io.shardingsphere.core.constant.SQLType;
 import io.shardingsphere.core.merger.MergeEngineFactory;
@@ -25,7 +24,6 @@ import io.shardingsphere.core.merger.MergedResult;
 import io.shardingsphere.core.merger.QueryResult;
 import io.shardingsphere.core.parsing.SQLJudgeEngine;
 import io.shardingsphere.core.parsing.parser.sql.SQLStatement;
-import io.shardingsphere.core.parsing.parser.sql.dml.insert.InsertStatement;
 import io.shardingsphere.core.routing.PreparedStatementRoutingEngine;
 import io.shardingsphere.core.routing.SQLExecutionUnit;
 import io.shardingsphere.core.routing.SQLRouteResult;
@@ -41,30 +39,32 @@ import io.shardingsphere.proxy.transport.mysql.packet.command.CommandResponsePac
 import io.shardingsphere.proxy.transport.mysql.packet.command.statement.PreparedStatementRegistry;
 import io.shardingsphere.proxy.transport.mysql.packet.command.statement.execute.BinaryResultSetRowPacket;
 import io.shardingsphere.proxy.transport.mysql.packet.command.statement.execute.PreparedStatementParameter;
-import io.shardingsphere.proxy.transport.mysql.packet.command.text.query.ColumnDefinition41Packet;
 import io.shardingsphere.proxy.transport.mysql.packet.command.text.query.FieldCountPacket;
 import io.shardingsphere.proxy.transport.mysql.packet.generic.EofPacket;
 import io.shardingsphere.proxy.transport.mysql.packet.generic.ErrPacket;
 import io.shardingsphere.proxy.transport.mysql.packet.generic.OKPacket;
+import lombok.Getter;
+import lombok.Setter;
 
-import javax.sql.DataSource;
-import javax.sql.rowset.CachedRowSet;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.Iterator;
-import java.util.LinkedList;
 import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Future;
 
 /**
  * Statement execute backend handler.
  *
  * @author zhangyonglun
  */
+@Getter
+@Setter
 public final class StatementExecuteBackendHandler implements BackendHandler {
     
     private static final Integer FETCH_ONE_ROW_A_TIME = Integer.MIN_VALUE;
@@ -95,9 +95,9 @@ public final class StatementExecuteBackendHandler implements BackendHandler {
     
     public StatementExecuteBackendHandler(final List<PreparedStatementParameter> preparedStatementParameters, final int statementId, final DatabaseType databaseType, final boolean showSQL) {
         this.preparedStatementParameters = preparedStatementParameters;
-        connections = new ArrayList<>(1024);
-        resultSets = new ArrayList<>(1024);
-        columnTypes = new ArrayList<>(32);
+        connections = new CopyOnWriteArrayList<>();
+        resultSets = new CopyOnWriteArrayList<>();
+        columnTypes = new CopyOnWriteArrayList<>();
         isMerged = false;
         hasMoreResultValueFlag = true;
         this.databaseType = databaseType;
@@ -122,9 +122,22 @@ public final class StatementExecuteBackendHandler implements BackendHandler {
         MasterSlaveRouter masterSlaveRouter = new MasterSlaveRouter(RuleRegistry.getInstance().getMasterSlaveRule());
         SQLStatement sqlStatement = new SQLJudgeEngine(sql).judge();
         String dataSourceName = masterSlaveRouter.route(sqlStatement.getType()).iterator().next();
-        List<CommandResponsePackets> result = new LinkedList<>();
-        result.add(execute(sqlStatement, dataSourceName, sql));
-        return merge(sqlStatement, result);
+        List<CommandResponsePackets> packets = new CopyOnWriteArrayList<>();
+        ExecutorService executorService = RuleRegistry.getInstance().getExecutorService();
+        List<Future<CommandResponsePackets>> resultList = new ArrayList<>(1024);
+        resultList.add(executorService.submit(new StatementExecuteWorker(this, sqlStatement, dataSourceName, sql)));
+        for (Future<CommandResponsePackets> each : resultList) {
+            try {
+                while (!each.isDone()) {
+                    continue;
+                }
+                packets.add(each.get());
+            } catch (InterruptedException | ExecutionException ex) {
+                return new CommandResponsePackets(new ErrPacket(1, 0, "", "", "" + ex.getMessage()));
+            
+            }
+        }
+        return merge(sqlStatement, packets);
     }
     
     private CommandResponsePackets executeForSharding() {
@@ -135,28 +148,26 @@ public final class StatementExecuteBackendHandler implements BackendHandler {
         if (routeResult.getExecutionUnits().isEmpty()) {
             return new CommandResponsePackets(new OKPacket(1, 0, 0, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue(), 0, ""));
         }
-        List<CommandResponsePackets> packets = new LinkedList<>();
+        List<CommandResponsePackets> packets = new CopyOnWriteArrayList<>();
+        ExecutorService executorService = RuleRegistry.getInstance().getExecutorService();
+        List<Future<CommandResponsePackets>> resultList = new ArrayList<>(1024);
         for (SQLExecutionUnit each : routeResult.getExecutionUnits()) {
-            // TODO multiple threads
-            packets.add(execute(routeResult.getSqlStatement(), each.getDataSource(), each.getSqlUnit().getSql()));
+            resultList.add(executorService.submit(new StatementExecuteWorker(this, routeResult.getSqlStatement(), each.getDataSource(), each.getSqlUnit().getSql())));
+        }
+        for (Future<CommandResponsePackets> each : resultList) {
+            try {
+                while (!each.isDone()) {
+                    continue;
+                }
+                packets.add(each.get());
+            } catch (InterruptedException | ExecutionException ex) {
+                return new CommandResponsePackets(new ErrPacket(1, 0, "", "", "" + ex.getMessage()));
+            
+            }
         }
         CommandResponsePackets result = merge(routeResult.getSqlStatement(), packets);
         ProxyShardingRefreshHandler.build(routeResult).execute();
         return result;
-    }
-    
-    private CommandResponsePackets execute(final SQLStatement sqlStatement, final String dataSourceName, final String sql) {
-        switch (sqlStatement.getType()) {
-            case DQL:
-            case DAL:
-                return executeQuery(RuleRegistry.getInstance().getDataSourceMap().get(dataSourceName), sql);
-            case DML:
-            case DDL:
-                return RuleRegistry.getInstance().isOnlyMasterSlave() ? executeUpdate(RuleRegistry.getInstance().getDataSourceMap().get(dataSourceName), sql)
-                    : executeUpdate(RuleRegistry.getInstance().getDataSourceMap().get(dataSourceName), sql, sqlStatement);
-            default:
-                return executeCommon(RuleRegistry.getInstance().getDataSourceMap().get(dataSourceName), sql);
-        }
     }
     
     private List<Object> getComStmtExecuteParameters() {
@@ -167,175 +178,10 @@ public final class StatementExecuteBackendHandler implements BackendHandler {
         return result;
     }
     
-    private void setJDBCPreparedStatementParameters(final PreparedStatement preparedStatement) throws SQLException {
+    public void setJDBCPreparedStatementParameters(final PreparedStatement preparedStatement) throws SQLException {
         for (int i = 0; i < getComStmtExecuteParameters().size(); i++) {
             preparedStatement.setObject(i + 1, getComStmtExecuteParameters().get(i));
         }
-    }
-    
-    private CommandResponsePackets executeQuery(final DataSource dataSource, final String sql) {
-        if (ProxyMode.MEMORY_STRICTLY == ProxyMode.valueOf(RuleRegistry.getInstance().getProxyMode())) {
-            return executeQueryWithStreamResultSet(dataSource, sql);
-        } else if (ProxyMode.CONNECTION_STRICTLY == ProxyMode.valueOf(RuleRegistry.getInstance().getProxyMode())) {
-            return executeQueryWithNonStreamResultSet(dataSource, sql);
-        } else {
-            return new CommandResponsePackets(new ErrPacket(1, 0, "", "", "Invalid proxy.mode"));
-        }
-    }
-    
-    private CommandResponsePackets executeQueryWithStreamResultSet(final DataSource dataSource, final String sql) {
-        try {
-            Connection connection = dataSource.getConnection();
-            connections.add(connection);
-            PreparedStatement preparedStatement = connection.prepareStatement(sql);
-            preparedStatement.setFetchSize(FETCH_ONE_ROW_A_TIME);
-            setJDBCPreparedStatementParameters(preparedStatement);
-            resultSets.add(preparedStatement.executeQuery());
-            return getQueryDatabaseProtocolPackets();
-        } catch (final SQLException ex) {
-            return new CommandResponsePackets(new ErrPacket(1, ex.getErrorCode(), "", ex.getSQLState(), ex.getMessage()));
-        }
-    }
-    
-    private CommandResponsePackets executeQueryWithNonStreamResultSet(final DataSource dataSource, final String sql) {
-        try (
-            Connection connection = dataSource.getConnection();
-            PreparedStatement preparedStatement = connection.prepareStatement(sql)
-        ) {
-            setJDBCPreparedStatementParameters(preparedStatement);
-            ResultSet resultSet = preparedStatement.executeQuery();
-            CachedRowSet cachedRowSet = new CachedRowSetImpl();
-            cachedRowSet.populate(resultSet);
-            resultSets.add(cachedRowSet);
-            return getQueryDatabaseProtocolPackets();
-        } catch (final SQLException ex) {
-            return new CommandResponsePackets(new ErrPacket(1, ex.getErrorCode(), "", ex.getSQLState(), ex.getMessage()));
-        }
-    }
-    
-    private CommandResponsePackets executeUpdate(final DataSource dataSource, final String sql, final SQLStatement sqlStatement) {
-        PreparedStatement preparedStatement = null;
-        try (
-            Connection connection = dataSource.getConnection()) {
-            int affectedRows;
-            long lastInsertId = 0;
-            if (sqlStatement instanceof InsertStatement) {
-                preparedStatement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS);
-                setJDBCPreparedStatementParameters(preparedStatement);
-                affectedRows = preparedStatement.executeUpdate();
-                lastInsertId = getGeneratedKey(preparedStatement);
-            } else {
-                preparedStatement = connection.prepareStatement(sql);
-                setJDBCPreparedStatementParameters(preparedStatement);
-                affectedRows = preparedStatement.executeUpdate();
-            }
-            return new CommandResponsePackets(new OKPacket(1, affectedRows, lastInsertId, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue(), 0, ""));
-        } catch (final SQLException ex) {
-            return new CommandResponsePackets(new ErrPacket(1, ex.getErrorCode(), "", ex.getSQLState(), ex.getMessage()));
-        } finally {
-            MasterVisitedManager.clear();
-            if (null != preparedStatement) {
-                try {
-                    preparedStatement.close();
-                } catch (final SQLException ignore) {
-                }
-            }
-        }
-    }
-    
-    private CommandResponsePackets executeUpdate(final DataSource dataSource, final String sql) {
-        try (Connection connection = dataSource.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
-            setJDBCPreparedStatementParameters(preparedStatement);
-            int affectedRows = preparedStatement.executeUpdate();
-            ResultSet resultSet = preparedStatement.getGeneratedKeys();
-            long lastInsertId = 0;
-            while (resultSet.next()) {
-                lastInsertId = resultSet.getLong(1);
-            }
-            return new CommandResponsePackets(new OKPacket(1, affectedRows, lastInsertId, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue(), 0, ""));
-        } catch (final SQLException ex) {
-            return new CommandResponsePackets(new ErrPacket(1, ex.getErrorCode(), "", ex.getSQLState(), ex.getMessage()));
-        } finally {
-            MasterVisitedManager.clear();
-        }
-    }
-    
-    private CommandResponsePackets executeCommon(final DataSource dataSource, final String sql) {
-        try (
-            Connection connection = dataSource.getConnection();
-            PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
-            setJDBCPreparedStatementParameters(preparedStatement);
-            boolean hasResultSet = preparedStatement.execute();
-            if (hasResultSet) {
-                return getCommonDatabaseProtocolPackets(preparedStatement.getResultSet());
-            } else {
-                return new CommandResponsePackets(new OKPacket(1, preparedStatement.getUpdateCount(), 0, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue(), 0, ""));
-            }
-        } catch (final SQLException ex) {
-            return new CommandResponsePackets(new ErrPacket(1, ex.getErrorCode(), "", ex.getSQLState(), ex.getMessage()));
-        } finally {
-            MasterVisitedManager.clear();
-        }
-    }
-    
-    private CommandResponsePackets getQueryDatabaseProtocolPackets() throws SQLException {
-        CommandResponsePackets result = new CommandResponsePackets();
-        int currentSequenceId = 0;
-        ResultSetMetaData resultSetMetaData = resultSets.get(resultSets.size() - 1).getMetaData();
-        columnCount = resultSetMetaData.getColumnCount();
-        if (0 == columnCount) {
-            result.addPacket(new OKPacket(++currentSequenceId, 0, 0, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue(), 0, ""));
-            return result;
-        }
-        result.addPacket(new FieldCountPacket(++currentSequenceId, columnCount));
-        for (int i = 1; i <= columnCount; i++) {
-            ColumnType columnType = ColumnType.valueOfJDBCType(resultSetMetaData.getColumnType(i));
-            ColumnDefinition41Packet columnDefinition41Packet = new ColumnDefinition41Packet(++currentSequenceId, resultSetMetaData.getSchemaName(i), resultSetMetaData.getTableName(i),
-                resultSetMetaData.getTableName(i), resultSetMetaData.getColumnLabel(i), resultSetMetaData.getColumnName(i), resultSetMetaData.getColumnDisplaySize(i), columnType, 0);
-            result.addPacket(columnDefinition41Packet);
-            columnTypes.add(columnType);
-        }
-        result.addPacket(new EofPacket(++currentSequenceId, 0, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue()));
-        return result;
-    }
-    
-    private CommandResponsePackets getCommonDatabaseProtocolPackets(final ResultSet resultSet) throws SQLException {
-        CommandResponsePackets result = new CommandResponsePackets();
-        int currentSequenceId = 0;
-        ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
-        columnCount = resultSetMetaData.getColumnCount();
-        if (0 == columnCount) {
-            result.addPacket(new OKPacket(++currentSequenceId, 0, 0, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue(), 0, ""));
-            return result;
-        }
-        result.addPacket(new FieldCountPacket(++currentSequenceId, columnCount));
-        for (int i = 1; i <= columnCount; i++) {
-            ColumnType columnType = ColumnType.valueOfJDBCType(resultSetMetaData.getColumnType(i));
-            ColumnDefinition41Packet columnDefinition41Packet = new ColumnDefinition41Packet(++currentSequenceId, resultSetMetaData.getSchemaName(i), resultSetMetaData.getTableName(i),
-                resultSetMetaData.getTableName(i), resultSetMetaData.getColumnLabel(i), resultSetMetaData.getColumnName(i), resultSetMetaData.getColumnDisplaySize(i), columnType, 0);
-            result.addPacket(columnDefinition41Packet);
-            columnTypes.add(columnType);
-        }
-        result.addPacket(new EofPacket(++currentSequenceId, 0, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue()));
-        while (resultSet.next()) {
-            List<Object> data = new ArrayList<>(columnCount);
-            for (int i = 1; i <= columnCount; i++) {
-                data.add(resultSet.getObject(i));
-            }
-            result.addPacket(new BinaryResultSetRowPacket(++currentSequenceId, columnCount, data, columnTypes));
-        }
-        result.addPacket(new EofPacket(++currentSequenceId, 0, StatusFlag.SERVER_STATUS_AUTOCOMMIT.getValue()));
-        return result;
-    }
-    
-    private long getGeneratedKey(final PreparedStatement preparedStatement) throws SQLException {
-        long result = 0;
-        ResultSet resultSet = preparedStatement.getGeneratedKeys();
-        if (resultSet.next()) {
-            result = resultSet.getLong(1);
-        }
-        return result;
     }
     
     private CommandResponsePackets merge(final SQLStatement sqlStatement, final List<CommandResponsePackets> packets) {
